requirements.txt
Django==4.0
django-cors-headers==3.13.0
django-jazzmin==2.6.0
django-environ==0.9.0
django-filter==22.1
django-rest-swagger==2.2.0
django-timezone-field==5.0
django-email-verification==0.3.1
djangorestframework==3.13.1
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.2.0
drf-spectacular==0.23.1
drf-spectacular-sidecar==2022.9.1
psycopg2-binary==2.9.3
Pillow==9.3.0